GITHUB#50 Add VITE_APP_HIDE_FORCE_REFRESH_HINT to allow disabling hint pop-in

+++ client/src/observer/showForceRefreshHint.js
@@ -7,15 +7,17 @@ const forceRefreshHintShownKey = config.VITE_APP_LOCALSTORAGE_KEY_PREFIX + 'forc
 
 export default {
   install() {
-    eventBus.$on('app-mounted', function () {
-      if (!window.localStorage.getItem(forceRefreshHintShownKey)) {
-        store.commit('dialog/showMessage', {
-          title: i18n.global.t('notice.force-refresh-hint-title'),
-          message: i18n.global.t('notice.force-refresh-hint-message'),
-          closeButtonLabel: i18n.global.t('ok-thanks')
-        })
-        window.localStorage.setItem(forceRefreshHintShownKey, true)
-      }
-    })
+    if (!config.VITE_APP_HIDE_FORCE_REFRESH_HINT) {
+      eventBus.$on('app-mounted', function () {
+        if (!window.localStorage.getItem(forceRefreshHintShownKey)) {
+          store.commit('dialog/showMessage', {
+            title: i18n.global.t('notice.force-refresh-hint-title'),
+            message: i18n.global.t('notice.force-refresh-hint-message'),
+            closeButtonLabel: i18n.global.t('ok-thanks')
+          })
+          window.localStorage.setItem(forceRefreshHintShownKey, true)
+        }
+      })
+    }
   }
 }
